requirements.txt
# Multi-Confirmation False Breakout Reversal v1.0
# Requirements - Frozen until 50 trades completed

# Core Data & Analysis
pandas==2.1.4
numpy==1.26.2

# Backtesting
backtrader==1.9.78.123

# IBKR Data Source
ib-insync==0.9.86

# Configuration & Utilities
pyyaml==6.0.1
python-dotenv==1.0.0

# Enhanced Logging
loguru==0.7.2

# Time handling
pytz==2023.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Optional: Data Analysis (for notebooks)
matplotlib==3.8.2
seaborn==0.13.0
jupyter==1.0.0
